Refactor network example
