+++ bee-network/examples/common/mod.rs
@@ -1,16 +1,10 @@
 // Copyright 2020 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-pub use utf8msg::Utf8Message;
-
 pub mod logger;
 
-mod args;
 mod backend;
-mod config;
 mod utf8msg;
 
-pub use args::*;
 pub use backend::*;
-pub use config::*;
 pub use utf8msg::*;
